requirements.txt
accelerate>=0.15.0
bitsandbytes
colossalai==0.2.5
diffusers>=0.12.1
safetensors
einops
hydra-core
loguru
matplotlib
numpy
omegaconf
opencv_python
Pillow
pyarrow
PyYAML
scikit_learn
torch
torchvision
tqdm
transformers==4.25.1
xformers
